original_implementation_files/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic~=2.12.3
pydantic-settings~=2.11.0
tenacity
boto3>=1.29.0
pillow>=10.3.0
minio~=7.2.7
opencv-python
langchain-community
langchain-openai
ollama
qdrant-client[fastembed]==1.13.0
sentence-transformers==2.2.2
redis==5.0.1
celery
PyYAML
pytest
pytest-asyncio
pytest-env
pydantic-ai
pydantic-ai-slim[openai]

# New dependencies for production upgrades
asyncpg>=0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.9  # PostgreSQL sync driver (backup)
cryptography>=42.0.0
slowapi>=0.1.9
docker>=7.0.0
